feat: solve comments, removing private functions
