+++ medmodels/data_synthesis/mtgan/model/real_gru/real_gru.pyi
@@ -1,13 +1,23 @@
 import torch
 from torch import nn
 
+from medmodels.data_synthesis.mtgan.model.gan import (
+    TrainingHyperparametersTotal,
+)
+from medmodels.data_synthesis.mtgan.model.loaders import MTGANDataLoader
+
 class RealGRU(nn.Module):
     number_codes: int
     gru: nn.GRU
-    linear: nn.Sequential
+    linear_layer: nn.Sequential
 
-    def __init__(self, number_codes: int, real_gru_hidden_dimension: int) -> None: ...
+    epochs: int
+
+    def __init__(
+        self, number_codes: int, hyperparamaters: TrainingHyperparametersTotal
+    ) -> None: ...
     def forward(self, data: torch.Tensor) -> torch.Tensor: ...
     def calculate_hidden(
         self, data: torch.Tensor, number_admissions: torch.Tensor
     ) -> torch.Tensor: ...
+    def train(self, train_loader: MTGANDataLoader) -> None: ...
