+++ medmodels/data_synthesis/mtgan/modules/preprocessor.pyi
@@ -1,4 +1,4 @@
-from typing import Dict, List, Tuple, TypedDict
+from typing import Dict, TypedDict
 
 from torch import nn
 
@@ -6,7 +6,6 @@ from medmodels.medrecord.medrecord import MedRecord
 from medmodels.medrecord.types import (
     Group,
     MedRecordAttribute,
-    MedRecordValue,
     NodeIndex,
 )
 
@@ -43,27 +42,4 @@ class MTGANPreprocessor(nn.Module):
         concepts_group: Group = "concepts",
         time_attribute: str = "time",
     ) -> None: ...
-    def _remove_uncommon_concepts(
-        self,
-        medrecord: MedRecord,
-        min_number_occurrences: int,
-        return_dict: bool = False,
-    ) -> Tuple[Dict[int, NodeIndex], MedRecord]: ...
-    def _find_first_admission(self, medrecord: MedRecord) -> MedRecord: ...
-    def _sample_patients(
-        self, medrecord: MedRecord, num_sampled_patients: int
-    ) -> MedRecord: ...
-    def _get_attribute(
-        self, medrecord: MedRecord, attribute_name: MedRecordAttribute
-    ) -> MedRecordAttribute: ...
-    def _remove_low_number_time_windows(
-        self,
-        medrecord: MedRecord,
-        patient_index: NodeIndex,
-        absolute_time_attribute: MedRecordAttribute,
-        min_codes_per_window: int,
-    ) -> List[MedRecordValue]: ...
-    def _find_relative_times(
-        self, medrecord: MedRecord, time_interval_days: int, min_codes_per_window: int
-    ) -> MedRecord: ...
     def preprocess(self, medrecord: MedRecord) -> MedRecord: ...
